Rename hallmonitor.py to app.py
