+++ scripts/monitor/hallMonitor2/hallmonitor/tests/integration/test_misc.py
@@ -285,7 +285,7 @@ class DataDictionaryHasChangesTestCase(TestCase):
         return modified_files
 
     def validate(self):
-        from hallmonitor.hallmonitor import main
+        from hallmonitor.app import main
 
         with pytest.raises(ValueError, match="Data dictionary has changed"):
             args = self.get_standard_args()
@@ -333,7 +333,7 @@ class PendingFilesCsvCreatedTestCase(TestCase):
         # Raw data validation deals with pending-files and pending-errors,
         # while checked data validation only uses pending-errors. In this
         # test case, we want to examine the program's behavior for both.
-        from hallmonitor.hallmonitor import raw_data_validation
+        from hallmonitor.app import raw_data_validation
 
         # save start time for later
         start_dt = datetime.datetime.now(tz=TZ_INFO)
@@ -500,10 +500,10 @@ class DeleteFailedCheckedIdentifierTestCase(TestCase):
         return modified_files
 
     def validate(self):
-        from hallmonitor import hallmonitor
+        from hallmonitor import app
 
         try:
-            hallmonitor.checked_data_validation(logger=Mock(), dataset=self.case_dir)
+            app.checked_data_validation(logger=Mock(), dataset=self.case_dir)
         except Exception as err:
             raise AssertionError from err
 
@@ -557,10 +557,10 @@ class KeepFailedIdentifierCheckedNoRawTestCase(TestCase):
         return modified_files
 
     def validate(self):
-        from hallmonitor import hallmonitor
+        from hallmonitor import app
 
         try:
-            hallmonitor.checked_data_validation(logger=Mock(), dataset=self.case_dir)
+            app.checked_data_validation(logger=Mock(), dataset=self.case_dir)
         except Exception as err:
             raise AssertionError from err
